Successfully searching for restaurants and placing markers in the desired radius. Still need to add InfoPopUps and Geolocation support for changing the address to search for food spots at

+++ client/src/GraphQL/localMutations.js
@@ -36,7 +36,7 @@ export const updateSearchParameters = gql`
 `;
 
 export const updateMapPosition = gql`
-  mutation updateMapPosition($latitude: Int!, $longitude: Int!) {
+  mutation updateMapPosition($latitude: Float!, $longitude: Float!) {
     updateMapPosition(latitude: $latitude, longitude: $longitude) @client {
       latitude
       longitude
